requirements.txt
django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=3.15
gunicorn>=21.2.0
scikit-learn>=1.3.0
pandas>=2.0
numpy>=1.25
joblib>=1.3.0
whitenoise>=6.5.0
